# examples/output-sqlite/openapi.yaml
...
openapi: 3.0.3
info:
  title: InventoryPlatform
  version: 1.0.0
  description: Generated by MarrowScript compiler
servers:
  - 
    url: "http://localhost:3000"
paths:
...
  /trade_offers:
    get:
      summary: List TradeOffer
      operationId: listTradeOffer
      tags:
        - TradeOffer
      parameters:
        - 
          name: page
          in: query
          schema:
            type: integer
            default: 1
        - 
          name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        200:
          description: List of TradeOffer
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TradeOffer"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        401:
          description: Unauthorized
    post:
      summary: Create TradeOffer
      operationId: createTradeOffer
      tags:
        - TradeOffer
      security:
        - 
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeOffer"
      responses:
        200:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeOffer"
        400:
          description: Bad request
        401:
          description: Unauthorized
        422:
          description: Precondition failed
  /trade_offers/{id}:
    get:
      summary: Get TradeOffer
      operationId: getTradeOffer
      tags:
        - TradeOffer
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      responses:
        200:
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeOffer"
        400:
          description: Not found
        401:
          description: Unauthorized
    put:
      summary: Update TradeOffer
      operationId: updateTradeOffer
      tags:
        - TradeOffer
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeOffer"
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeOffer"
        400:
          description: Bad request
        401:
          description: Unauthorized
        422:
          description: Precondition failed
    delete:
      summary: Delete TradeOffer
      operationId: deleteTradeOffer
      tags:
        - TradeOffer
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      responses:
        200:
          description: Deleted
        400:
          description: Not found
        401:
          description: Unauthorized
  /trade_offers/accept-trade:
    post:
      summary: accept_trade on TradeOffer
      operationId: accept_tradeTradeOffer
      tags:
        - TradeOffer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                trade_id:
                  type: string
                  format: uuid
                  description: UUID of an existing TradeOffer row (alternative to inline body)
                id:
                  type: string
                  format: uuid
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
                from_player:
                  type: string
                  format: uuid
                to_player:
                  type: string
                  format: uuid
                offered_items:
                  type: array
                  items:
                    type: string
                    format: uuid
                requested_items:
                  type: array
                  items:
                    type: string
                    format: uuid
                status:
                  type: string
                expires_at:
                  type: string
                  format: date-time
                acceptor_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Player row (alternative to inline body)
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  action:
                    type: string
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        422:
          description: Precondition failed
      security:
        - 
          BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TradeOffer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        from_player:
          type: string
          format: uuid
        to_player:
          type: string
          format: uuid
        offered_items:
          type: array
          items:
            type: string
            format: uuid
        requested_items:
          type: array
          items:
            type: string
            format: uuid
        status:
          type: string
        expires_at:
          type: string
          format: date-time