# examples/output-sqlite/src/routes/trade_offer.ts
// Generated by MarrowScript compiler. DO NOT EDIT.
// Service: TradeOfferService
import { Router, Request, Response } from "express";
import { v4 as uuid } from "uuid";
import { query, queryOne, execute, pool } from "../db";
import { eventBus } from "../events";
import { requireAuth, AuthContext } from "../auth";
import rateLimit from "express-rate-limit";
import { auditLog } from "../audit";
import { logger } from "../logger";
import { counter } from "../metrics";
import { TradeOfferCreateSchema, TradeOfferUpdateSchema } from "../schemas";
import * as __algorithms from "../algorithms";
const { shortestPath, topologicalSort, binarySearch, bipartiteMatching, roundRobin, weightedAverage, percentile, rankBy, consistentHash } = __algorithms as any;

import { transitionTradeOffer, TRADEOFFER_INITIAL } from "../state_machines/trade_offer";

export const tradeOffersRouter = Router();

// Rate limiter from policy declaration
const __routeRateLimit = rateLimit({ windowMs: 60000, max: 100, standardHeaders: true, legacyHeaders: false });

// CREATE
tradeOffersRouter.post("/", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  // Validate the request body against the generated Zod schema.
  const __parsed = TradeOfferCreateSchema.safeParse(req.body);
  if (!__parsed.success) {
    return res.status(400).json({ error: { code: "VALIDATION_FAILED", message: "Invalid request body", issues: __parsed.error.flatten() } });
  }
  const __body = __parsed.data;
  try {
    const id = uuid();
    const { from_player, to_player, offered_items, requested_items, status, expires_at } = __body as any;
    const state = TRADEOFFER_INITIAL;
    const sql = `INSERT INTO trade_offers (id, from_player, to_player, offered_items, requested_items, status, expires_at, state) VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING *`;
    const rows = await query(sql, [id, from_player, to_player, offered_items, requested_items, status, expires_at, state]);
    counter("entity.created", { entity: "TradeOffer" });
    res.status(201).json(rows[0]);
  } catch (e: any) {
    res.status(400).json({ error: { code: "CREATE_FAILED", message: e.message } });
  }
});

// READ
tradeOffersRouter.get("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const row = await queryOne(`SELECT * FROM trade_offers WHERE id = $1`, [req.params.id]);
    if (!row) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    res.json(row);
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// LIST
tradeOffersRouter.get("/", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const page = parseInt(req.query.page as string) || 1;
    const pageSize = Math.min(parseInt(req.query.page_size as string) || 50, 100);
    const offset = (page - 1) * pageSize;
    const rows = await query(`SELECT * FROM trade_offers ORDER BY created_at DESC LIMIT $1 OFFSET $2`, [pageSize, offset]);
    const [{ count }] = await query(`SELECT COUNT(*) as count FROM trade_offers`);
    res.json({ items: rows, total: parseInt(count), page, page_size: pageSize });
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// UPDATE
const __tradeOffersUpdatable = new Set<string>(["from_player","to_player","offered_items","requested_items","status","expires_at","state"]);
tradeOffersRouter.put("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  // 1. Validate body shape and types via Zod (UpdateSchema is partial).
  const __parsed = TradeOfferUpdateSchema.safeParse(req.body);
  if (!__parsed.success) {
    return res.status(400).json({ error: { code: "VALIDATION_FAILED", message: "Invalid request body", issues: __parsed.error.flatten() } });
  }
  // 2. Reject unknown columns (defense in depth — Zod would already strip them).
  const __unknown = Object.keys(req.body || {}).filter(k => !__tradeOffersUpdatable.has(k));
  if (__unknown.length > 0) {
    return res.status(400).json({ error: { code: "UNKNOWN_FIELDS", message: `Unknown fields: ${__unknown.join(", ")}`, fields: __unknown } });
  }
  // 3. Use the Zod-parsed object as the update set.
  const fields: Record<string, unknown> = __parsed.data as Record<string, unknown>;
  if (Object.keys(fields).length === 0) {
    return res.status(400).json({ error: { code: "NO_FIELDS", message: "No updatable fields supplied" } });
  }
  // State machine enforcement
  if (fields.state !== undefined) {
    const current = await queryOne<{ state: string }>(`SELECT state FROM trade_offers WHERE id = $1`, [req.params.id]);
    if (!current) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    // Find the trigger for this state transition
    const trigger = `${current.state}_to_${fields.state}`;
    const tr = transitionTradeOffer(current.state as any, trigger);
    if (!tr.ok) return res.status(422).json({ error: { code: "INVALID_TRANSITION", message: `Cannot transition from ${current.state} to ${fields.state}` } });
  }
  const keys = Object.keys(fields);
  const sets = keys.map((k, i) => `${k} = $${i + 2}`).join(", ");
  const values = keys.map(k => fields[k]);
  const sql = `UPDATE trade_offers SET ${sets}, updated_at = NOW() WHERE id = $1 RETURNING *`;
  const rows = await query(sql, [req.params.id, ...values]);
  if (rows.length === 0) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
  res.json(rows[0]);
});

// DELETE
tradeOffersRouter.delete("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const count = await execute(`DELETE FROM trade_offers WHERE id = $1`, [req.params.id]);
    if (count === 0) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    res.status(204).send();
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// CAPABILITY: accept_trade [transactional]
tradeOffersRouter.post("/accept-trade", __routeRateLimit, requireAuth, auditLog("accept_trade", "TradeOffer"), async (req: Request, res: Response) => {
  const auth: AuthContext = (req as any).auth;
  const __client = await pool.connect();
  try {
    await __client.query("BEGIN");
    // Fetch entities
    const trade = await queryOne(`SELECT * FROM trade_offers WHERE id = $1`, [req.body.trade_id || req.params.id]);
    if (!trade) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "trade not found" } });
    }
    const acceptor = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.body.acceptor_id || req.params.id]);
    if (!acceptor) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "acceptor not found" } });
    }

    // Preconditions
    if (!(trade?.status === "pending")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "trade.status == \\\"pending\\\"" } });
    }
    if (!(trade?.to_player === acceptor?.id)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "trade.to_player == acceptor.id" } });
    }
    if (!(acceptor?.state === "active")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "acceptor.state == \\\"active\\\"" } });
    }
    if (!(trade?.expires_at > new Date())) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "trade.expires_at > now()" } });
    }

    // Effects (applied in declaration order)
    // EFFECT: trade.status = "accepted"
    // TODO: Implement this effect manually
    // EFFECT: trade.offered_items.owner_id = trade.to_player
    // TODO: Implement this effect manually
    // EFFECT: trade.requested_items.owner_id = trade.from_player
    // TODO: Implement this effect manually

    // Emit events
    await eventBus.publish("ItemTraded", { trade_id: trade?.id, acceptor_id: acceptor?.id, timestamp: new Date().toISOString(), actor_id: auth.actor_id }, "TradeOfferService", auth.trace_id, __client);

    res.json({ ok: true, action: "accept_trade", entity: trade });
    await __client.query("COMMIT");
  } catch (e: any) {
    await __client.query("ROLLBACK");
    if (!res.headersSent) {
      res.status(400).json({ error: { code: "CAPABILITY_FAILED", message: e.message } });
    } else {
      logger.error("late_error_after_response", { event: "accept_trade", status: "rejected", metadata: { error: String(e?.message || e).slice(0, 256) } });
    }
  } finally {
    __client.release();
  }
});